Open spot card in dialog
