+++ src/frontend/actionTypes/index.js
@@ -47,6 +47,8 @@ export const CLOSE_DELETE_MAP_DIALOG = 'CLOSE_DELETE_MAP_DIALOG';
 export const SELECT_SPOT = 'SELECT_SPOT';
 export const OPEN_SPOT_CARD = 'OPEN_SPOT_CARD';
 export const CLOSE_SPOT_CARD = 'CLOSE_SPOT_CARD';
+export const OPEN_SPOT_DIALOG = 'OPEN_SPOT_DIALOG';
+export const CLOSE_SPOT_DIALOG = 'CLOSE_SPOT_DIALOG';
 export const FETCH_SPOT = 'FETCH_SPOT';
 export const FETCH_SPOT_REVIEWS = 'FETCH_SPOT_REVIEWS';
 export const FETCH_COLLABORATORS = 'FETCH_COLLABORATORS';
