+++ src/frontend/actions/closeSpotDialog.js
@@ -1,0 +1,9 @@
+import { CLOSE_SPOT_DIALOG } from '../actionTypes';
+
+const closeSpotDialog = () => {
+  return {
+    type: CLOSE_SPOT_DIALOG
+  };
+};
+
+export default closeSpotDialog;
