+++ src/frontend/components/Routes.js
@@ -4,12 +4,18 @@ import loadable from '@loadable/component';
 import createHistory from 'history/createBrowserHistory';
 import pathToRegexp from 'path-to-regexp';
 
-import { OPEN_REVIEW_DIALOG, SWITCH_SUMMARY } from '../actionTypes';
+import {
+  OPEN_REVIEW_DIALOG,
+  SWITCH_SUMMARY,
+  OPEN_SPOT_DIALOG
+} from '../actionTypes';
 
 import locationChange from '../actions/locationChange';
 import getHistory from '../actions/getHistory';
 import openReviewDialog from '../actions/openReviewDialog';
 import closeReviewDialog from '../actions/closeReviewDialog';
+import openSpotDialog from '../actions/openSpotDialog';
+import closeSpotDialog from '../actions/closeSpotDialog';
 import selectSpot from '../actions/selectSpot';
 import requestMapCenter from '../actions/requestMapCenter';
 import switchSummary from '../actions/switchSummary';
@@ -97,6 +103,11 @@ const modalRoutes = [
     path: '/maps/:mapId',
     action: SWITCH_SUMMARY,
     modal: true
+  },
+  {
+    path: '/spots/:placeId',
+    action: OPEN_SPOT_DIALOG,
+    modal: true
   }
 ];
 
@@ -139,6 +150,10 @@ const Routes = () => {
       case SWITCH_SUMMARY:
         dispatch(switchSummary());
         dispatch(closeReviewDialog());
+        break;
+      case OPEN_SPOT_DIALOG:
+        const spot = currentLocation.state.spot;
+        dispatch(openSpotDialog(spot));
         break;
       default:
         return;
