+++ src/frontend/locales/ja.json
@@ -144,6 +144,7 @@
   "create review success": "レポートを作成しました！",
   "edit review success": "レポートを更新しました！",
   "spot": "スポット",
+  "open in google maps": "Google マップで開く",
   "unable to get permission": "プッシュ通知の許可が確認できませんでした。",
   "unable to get registration token": "プッシュ通知が有効化できませんでした。",
   "push successfully enabled": "プッシュ通知が有効化されました。",
