+++ src/frontend/reducers/spotDetailReducer.js
@@ -4,12 +4,16 @@ import {
   CLEAR_SPOT_STATE,
   CREATE_REVIEW,
   EDIT_REVIEW,
-  DELETE_REVIEW
+  DELETE_REVIEW,
+  OPEN_SPOT_DIALOG,
+  CLOSE_SPOT_DIALOG,
+  LOCATION_CHANGE
 } from '../actionTypes';
 
 const initialState = {
   currentSpot: undefined,
-  spotReviews: []
+  spotReviews: [],
+  spotDialogOpen: false
 };
 
 const reducer = (state = initialState, action) => {
@@ -62,6 +66,19 @@ const reducer = (state = initialState, action) => {
         spotLoading: false,
         currentSpot: undefined,
         spotReviews: []
+      });
+    case OPEN_SPOT_DIALOG:
+      return Object.assign({}, state, {
+        spotDialogOpen: true,
+        currentSpot: action.payload.spot
+      });
+    case CLOSE_SPOT_DIALOG:
+      return Object.assign({}, state, {
+        spotDialogOpen: false
+      });
+    case LOCATION_CHANGE:
+      return Object.assign({}, state, {
+        spotDialogOpen: false
       });
     default:
       return state;
